// src/modern/shared/enums/membership.enum.ts
export enum MembershipState {
    ACTIVE = 'active',
    PENDING = 'pending',
    EXPIRED = 'expired'
}

export enum MembershipPeriodState {
    PLANNED = 'planned',
    ACTIVE = 'active',
    COMPLETED = 'completed'
}

export enum BillingInterval {
    WEEKLY = 'weekly',
    MONTHLY = 'monthly',
    YEARLY = 'yearly'
}

export enum PaymentMethod {
    CASH = 'cash',
    CARD = 'card'
} 